feat: codebase overhaul
