+++ pkg/eve/match/games/common.go
