+++ pkg/eve/tournament/openings.go
