+++ pkg/eve/tournament/schedule/round_robin.go
@@ -1,26 +1,6 @@
-package tournament
+package schedule
 
-import (
-	"fmt"
-	"slices"
-)
-
-func GetScheduler(name string) (Scheduler, error) {
-	switch name {
-	case "round-robin", "":
-		return &RoundRobin{}, nil
-	case "gauntlet":
-		return &Gauntlet{}, nil
-	default:
-		return nil, fmt.Errorf("new tour: invalid scheduler %s", name)
-	}
-}
-
-type Scheduler interface {
-	Initialize(int)
-	NextEncounter() (int, int)
-	TotalEncounters() int
-}
+import "slices"
 
 type RoundRobin struct {
 	player_count int
@@ -74,23 +54,4 @@ func (rr *RoundRobin) NextEncounter() (int, int) {
 
 func (rr *RoundRobin) TotalEncounters() int {
 	return rr.player_count * (rr.player_count - 1) / 2
-}
-
-type Gauntlet struct {
-	player_count int
-	game_number  int
-}
-
-func (g *Gauntlet) Initialize(n int) {
-	g.player_count = n
-	g.game_number = 0
-}
-
-func (g *Gauntlet) NextEncounter() (int, int) {
-	g.game_number++
-	return 0, g.game_number
-}
-
-func (g *Gauntlet) TotalEncounters() int {
-	return g.player_count - 1
 }
